fixup(rebase): cover all WelcomeView seeds + split artefacts file
Two adjustments after rebasing onto main, which had merged the
WelcomeView state-machine and three new `welcome*CloudProfile` seeds:

- `applySeedProgressFixtures` switch now exhaustively handles the
  three Welcome seeds added on main (no-op for them).
- Split `MoolahUITestCase`'s artefact-collection helpers into
  `MoolahUITestCase+Artefacts.swift`. The class body had crossed
  SwiftLint's 250-line `type_body_length` limit by one; extracting
  the artefact code into its own extension is the cohesive fix.
