+++ MoolahUITests_macOS/Helpers/MoolahUITestCase.swift
@@ -123,223 +123,4 @@ class MoolahUITestCase: XCTestCase {
     app.application.typeKey(key, modifierFlags: modifiers)
   }
 
-  // MARK: - Internal: artefact collection
-
-  private func collectFailureArtefacts(for app: MoolahApp, succeeded: Bool) {
-    if succeeded { return }
-
-    let dir = artefactDirectory(for: name)
-    do {
-      try FileManager.default.createDirectory(
-        at: dir, withIntermediateDirectories: true)
-    } catch {
-      // Recording the failure to attach is itself best-effort; keep going.
-      add(XCTAttachment(string: "Failed to create artefact dir: \(error)"))
-      return
-    }
-
-    // Print the path so `scripts/test-ui.sh` can grep it out and copy the
-    // artefacts back into `.agent-tmp/`, and attach an XCTAttachment so the
-    // test navigator surfaces it.
-    print("[MoolahUITestCase] ARTEFACT_DIR \(dir.path)")
-    add(XCTAttachment(string: "ui-fail artefacts written to: \(dir.path)\n"))
-
-    write(treeText(for: app), to: dir.appendingPathComponent("tree.txt"))
-    write(seedText(for: app), to: dir.appendingPathComponent("seed.txt"))
-    write(Trace.render(succeeded: false), to: dir.appendingPathComponent("trace.txt"))
-    captureScreenshot(for: app, to: dir.appendingPathComponent("screenshot.png"))
-
-    attachArtefacts(in: dir)
-  }
-
-  /// Resolves `<runner tmpdir>/MoolahUITests/ui-fail-<TestName>/`.
-  ///
-  /// The XCUITest runner runs in a stricter sandbox than the app: writes
-  /// to `/private/tmp/` fail with "Operation not permitted" on current
-  /// macOS. `FileManager.default.temporaryDirectory` resolves to the
-  /// runner's own per-process tmpdir under `/var/folders/.../T/` which
-  /// the runner can freely create inside and the developer's shell can
-  /// still read (same uid) without TCC prompts. `scripts/test-ui.sh`
-  /// greps the `[MoolahUITestCase] ARTEFACT_DIR` lines out of
-  /// xcodebuild's stdout and copies each artefact dir back into
-  /// `<repo-root>/.agent-tmp/` after `xcodebuild test` exits.
-  private func artefactDirectory(for testName: String) -> URL {
-    let cleanName =
-      testName
-      .replacingOccurrences(of: "-", with: "_")
-      .replacingOccurrences(of: "[", with: "")
-      .replacingOccurrences(of: "]", with: "")
-      .replacingOccurrences(of: " ", with: "_")
-    return FileManager.default.temporaryDirectory
-      .appendingPathComponent("MoolahUITests", isDirectory: true)
-      .appendingPathComponent("ui-fail-\(cleanName)", isDirectory: true)
-  }
-
-  // MARK: - Artefact contents
-
-  private func treeText(for app: MoolahApp) -> String {
-    var lines: [String] = []
-    lines.append("# accessibility tree (identifier | type | label | value | frame)")
-    if let focusedIdentifier = currentFocusedIdentifier(in: app) {
-      lines.append("# focused element: \(focusedIdentifier)")
-    } else {
-      lines.append("# focused element: (none — no element has keyboard focus)")
-    }
-    lines.append(
-      "# (per-element focus is omitted — `XCUIElementSnapshot` does not expose it.)")
-    lines.append("")
-    let snapshot = try? app.application.snapshot()
-    if let snapshot { appendTreeSnapshot(snapshot: snapshot, depth: 0, into: &lines) }
-    return lines.joined(separator: "\n") + "\n"
-  }
-
-  /// Walks the live element tree once to find the currently focused
-  /// element, returning its identifier (or label / type if no identifier
-  /// is set). At most one element has keyboard focus at a time.
-  private func currentFocusedIdentifier(in app: MoolahApp) -> String? {
-    let elements = app.application.descendants(matching: .any).allElementsBoundByIndex
-    for element in elements where (element.value(forKey: "hasKeyboardFocus") as? Bool) ?? false {
-      if !element.identifier.isEmpty { return element.identifier }
-      if !element.label.isEmpty { return "(label: \(element.label))" }
-      return "(\(element.elementType))"
-    }
-    return nil
-  }
-
-  private func appendTreeSnapshot(
-    snapshot: XCUIElementSnapshot, depth: Int, into lines: inout [String]
-  ) {
-    let indent = String(repeating: "  ", count: depth)
-    let identifier = snapshot.identifier.isEmpty ? "—" : snapshot.identifier
-    let type = String(describing: snapshot.elementType)
-    let label =
-      snapshot.label.isEmpty ? "—" : snapshot.label.replacingOccurrences(of: "\n", with: " ")
-    let value = (snapshot.value as? String).map { $0.isEmpty ? "—" : $0 } ?? "—"
-    let frame = snapshot.frame
-    let frameStr =
-      "(\(Int(frame.origin.x)),\(Int(frame.origin.y)) \(Int(frame.size.width))x\(Int(frame.size.height)))"
-    lines.append(
-      "\(indent)\(identifier) | \(type) | \(label) | \(value) | \(frameStr)"
-    )
-    for child in snapshot.children {
-      appendTreeSnapshot(snapshot: child, depth: depth + 1, into: &lines)
-    }
-  }
-
-  private func seedText(for app: MoolahApp) -> String {
-    var lines: [String] = []
-    lines.append("seed: \(app.seed.rawValue)")
-    lines.append("")
-    switch app.seed {
-    case .tradeBaseline:
-      appendTradeBaselineFixtures(into: &lines)
-    case .welcomeEmpty:
-      lines.append("# fixtures — empty index, no profiles")
-    case .welcomeSingleCloudProfile:
-      appendWelcomeSingleProfileFixtures(into: &lines)
-    case .welcomeMultipleCloudProfiles:
-      appendWelcomeMultipleProfileFixtures(into: &lines)
-    case .welcomeDownloading:
-      lines.append("# SyncProgress driven to .receiving with recordsReceivedThisSession=1234")
-      lines.append("# WelcomeView resolves to .heroDownloading(count: 1234)")
-    case .sidebarFooterUpToDate:
-      lines.append("# SyncProgress driven to .upToDate, lastSettledAt ~5 minutes ago")
-    case .sidebarFooterReceiving:
-      lines.append("# SyncProgress driven to .receiving with recordsReceivedThisSession=1234")
-    case .sidebarFooterSending:
-      lines.append("# SyncProgress driven to .upToDate with pendingUploads=12 then settled")
-    }
-    return lines.joined(separator: "\n") + "\n"
-  }
-
-  private func appendTradeBaselineFixtures(into lines: inout [String]) {
-    let fixtures = UITestFixtures.TradeBaseline.self
-    lines.append("# fixtures")
-    lines.append("profile.id      = \(fixtures.profileId)")
-    lines.append("profile.label   = \(fixtures.profileLabel)")
-    lines.append("profile.currency = \(fixtures.profileCurrencyCode)")
-    lines.append("checking.id     = \(fixtures.checkingAccountId)")
-    lines.append("checking.name   = \(fixtures.checkingAccountName)")
-    lines.append("brokerage.id    = \(fixtures.brokerageAccountId)")
-    lines.append("brokerage.name  = \(fixtures.brokerageAccountName)")
-    lines.append("usdSavings.id   = \(fixtures.usdAccountId)")
-    lines.append("usdSavings.name = \(fixtures.usdAccountName)")
-    lines.append("usdSavings.instrument = \(fixtures.usdAccountInstrumentCode)")
-    lines.append("trade.id        = \(fixtures.bhpPurchaseId)")
-    lines.append("trade.payee     = \(fixtures.bhpPurchasePayee)")
-    lines.append("trade.cents     = \(fixtures.bhpPurchaseAmountCents)")
-    lines.append("trade.date      = \(fixtures.bhpPurchaseDate)")
-    lines.append("historical.amount.cents = \(fixtures.historicalExpenseAmountCents)")
-    for (index, historical) in fixtures.historicalPayees.enumerated() {
-      lines.append(
-        "historical[\(index)].id/payee/date = "
-          + "\(historical.id) / \(historical.payee) / \(historical.date)"
-      )
-    }
-    lines.append(
-      "category.groceries.id/name = "
-        + "\(fixtures.groceriesCategoryId) / \(fixtures.groceriesCategoryName)")
-    lines.append("category.gym.id/name = \(fixtures.gymCategoryId) / \(fixtures.gymCategoryName)")
-    lines.append(
-      "splitShop.id/payee/date = "
-        + "\(fixtures.splitShopId) / \(fixtures.splitShopPayee) / \(fixtures.splitShopDate)")
-    lines.append(
-      "splitShop.legA.cents / legB.cents = "
-        + "\(fixtures.splitShopLegAAmountCents) / \(fixtures.splitShopLegBAmountCents)"
-    )
-  }
-
-  private func appendWelcomeSingleProfileFixtures(into lines: inout [String]) {
-    lines.append("# fixtures")
-    lines.append(
-      "household.id/label = "
-        + "\(UITestWelcomeFixtures.householdProfileId) "
-        + "/ \(UITestWelcomeFixtures.householdProfileLabel)"
-    )
-  }
-
-  private func appendWelcomeMultipleProfileFixtures(into lines: inout [String]) {
-    lines.append("# fixtures")
-    lines.append(
-      "household.id/label = "
-        + "\(UITestWelcomeFixtures.householdProfileId) "
-        + "/ \(UITestWelcomeFixtures.householdProfileLabel)"
-    )
-    lines.append(
-      "sideBusiness.id/label = "
-        + "\(UITestWelcomeFixtures.sideBusinessProfileId) "
-        + "/ \(UITestWelcomeFixtures.sideBusinessProfileLabel)"
-    )
-  }
-
-  private func captureScreenshot(for app: MoolahApp, to url: URL) {
-    let screenshot = app.application.screenshot()
-    do {
-      try screenshot.pngRepresentation.write(to: url)
-    } catch {
-      add(XCTAttachment(string: "Failed to write screenshot: \(error)"))
-    }
-  }
-
-  private func write(_ text: String, to url: URL) {
-    do {
-      try text.write(to: url, atomically: true, encoding: .utf8)
-    } catch {
-      add(XCTAttachment(string: "Failed to write \(url.lastPathComponent): \(error)"))
-    }
-  }
-
-  private func attachArtefacts(in dir: URL) {
-    let fileManager = FileManager.default
-    guard
-      let entries = try? fileManager.contentsOfDirectory(
-        at: dir, includingPropertiesForKeys: nil)
-    else { return }
-    for url in entries {
-      let attachment = XCTAttachment(contentsOfFile: url)
-      attachment.lifetime = .keepAlways
-      attachment.name = url.lastPathComponent
-      add(attachment)
-    }
-  }
 }
